[INLONG-123] plugin to common
