+++ inlong-agent/agent-common/src/main/java/org/apache/inlong/agent/dto/TaskRequestDto.java
@@ -15,29 +15,16 @@
  * limitations under the License.
  */
 
-package org.apache.inlong.agent.plugin.fetcher.dtos;
+package org.apache.inlong.agent.dto;
 
+import java.util.ArrayList;
+import java.util.List;
 import lombok.Data;
+import org.apache.inlong.agent.db.CommandEntity;
 
 @Data
-public class DataConfig {
-    private String additionalAttr;
-    private String inlongGroupId;
-    private String dataName;
-    private String inlongStreamId;
-    private String deliveryTime;
-    private String fieldSplitter;
-    private String cycleUnit;
-    private String ip;
-    private String middlewareType;
-    private String mqMasterAddress;
-    private String op;
-    private String scheduleTime;
-    private Integer taskId;
-    private String timeOffset;
-    private String topic;
-
-    public boolean isValid() {
-        return true;
-    }
+public class TaskRequestDto {
+    private String agentIp;
+    private String uuid;
+    private List<CommandEntity> commandInfo = new ArrayList<>();
 }
